improve perceived conversation load time
