+++ components/Conversation/RecipientControl.tsx
@@ -105,7 +105,7 @@ const RecipientControl = ({
   }
 
   return (
-    <div className="flex-1 flex-col flex h-14 bg-zinc-50 md:border md:border-gray-200 md:rounded-lg md:px-4 md:mx-4 md:mt-4 md:pb-1">
+    <div className="flex-1 flex-col flex h-14 bg-zinc-50 md:border md:border-gray-200 md:rounded-lg md:px-4 md:mx-4 md:mt-4 md:pb-1 md:mb-2">
       <form
         className="w-full flex pl-2 md:pl-0 h-full pt-1"
         action="#"
