+++ components/Conversation/index.ts
@@ -1,3 +1,4 @@
 export { default as RecipientControl } from './RecipientControl'
 export { default as MessagesList } from './MessagesList'
 export { default as MessageComposer } from './MessageComposer'
+export { default as Conversation } from './Conversation'
