+++ styles/MessageComposer.module.scss
@@ -1,7 +1,3 @@
-.container {
-  padding: 10px 0 10px 16px;
-}
-
 .bubble {
   border-radius: 24px;
   background-color: #fafafa;
